Set up a workspace, fixed more new-XCode-isms

+++ Set Application/SetCard.m
@@ -218,7 +218,7 @@ static void drawSquiggle(NSRect rect);
 	return FALSE;
 }
 
-- (unsigned int) hash {
+- (NSUInteger) hash {
 	return (
 			   shape
 			+ (color  << 8)
